Various updates
Game class now contains the objects in the scenes
Added an UI layer
Added a collider variable
